Rename image folders to `assets`
