+++ src/common/components/assets/OrbitalLogo.jpeg
